STITCH-2269 resolve linter warnings in StitchCoreSDK
